parser_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
hypercorn==0.14.4
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
playwright==1.40.0
websockets==12.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dotenv==1.0.0

python-multipart==0.0.6

easyocr==1.7.2
pillow==10.4.0
pytesseract==0.3.10
pymupdf==1.24.10

